描画順制御のためにwとhをIRenderReserveに増やした

+++ CImage.h
@@ -22,13 +22,13 @@ public:
 		}
 	}
 	
-	void Draw(int x1, int y1, double priority = 0.0, int num = 0);
-	void Draw(float x1, float y1, double priority = 0.0, int num = 0);
-	void DrawRota(int x1, int y1, float angle, float extend, double priority = 0.0, int num = 0);
-	void DrawRotaF(float x1, float y1, float angle, float extend, double priority = 0.0, int num = 0);
-	void DrawRotaFwithBlend(float x1, float y1, float angle, float extend, int color, int Blendmode, int value, double priority = 0.0, int num = 0);
-	void DrawExtendWithBlend(float x1, float y1, float x2, float y2, int color, int Blendmode, int value, double priority = 0.0, int num = 0);
-	void DrawCircleGauge(int x, int y, double ratio, double priority = 0.0, int num = 0);
+	void Draw(int x1, int y1, int priority = 0.0, int num = 0);
+	void Draw(float x1, float y1, int priority = 0.0, int num = 0);
+	void DrawRota(int x1, int y1, float angle, float extend, int priority = 0.0, int num = 0);
+	void DrawRotaF(float x1, float y1, float angle, float extend, int priority = 0.0, int num = 0);
+	void DrawRotaFwithBlend(float x1, float y1, float angle, float extend, int color, int Blendmode, int value, int priority = 0.0, int num = 0);
+	void DrawExtendWithBlend(float x1, float y1, float x2, float y2, int color, int Blendmode, int value, int priority = 0.0, int num = 0);
+	void DrawCircleGauge(int x, int y, double ratio, int priority = 0.0, int num = 0);
 	//void DrawOnCamela(int x, int y, int num = 0);		//ÉJÉÅÉâç¿ïWÇå≥Ç…ï`âÊÇ∑ÇÈä÷êîëΩï™óvÇÈÇÊÇÀ
 };
